refactor(Staff): Doing the pages server side to handle the admin role in the staff module
