+++ frontend/src/app/admin/staff/new/page.jsx
@@ -1,24 +1,18 @@
-"use client";
-import EmployeeForm from "../../../forms/EmployeeForm";
-import { createEmployee } from "../../../apiHandlers/adminStaff";
 import Layout from "@/app/components/base_layout/Layout";
+import { isAdmin } from "@/app/utils/requestBuilder";
+import NewEmployee from "./client";
 
-export default function NewEmployee() {
+export default async function Page() {
 
-  const submit = async(data) => {
-    try{
-      await createEmployee(data);
-    }catch(err){
-      return err;
-    }
-  };
+  const isAdm = await isAdmin();
   return (
     <Layout
       headerTitle={"Nuevo empleado"}
       mainTitle={"Registrar nuevo empleado"}
+      isAdmin={isAdm}
     >
       <title>SG BarberShop - Registrar Empleado</title>
-      <EmployeeForm onSubmit={submit} />
+      <NewEmployee />
     </Layout>
   );
 }
